FastApi/pyproject.toml
[tool.poetry]
name = "FastApi"
version = "0.1.0"
description = "FastApi with Starlette admin"
authors = ["Joshua Brown <Joshua.Brown@optix.co>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.27.0"
pydantic = "^2.5.3"
SQLAlchemy = "^2.0.25"
postgresql-access = "^1.0"
starlette-admin = "^0.13.1"
fastapi-common = "^0.3.12"
async-api-client = "^0.0.4"
async = "^0.6.2"
dotenv = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
